completed acc_heads

+++ app/Http/Controllers/API/PurchaseInvoiceController.php
@@ -5,9 +5,11 @@ namespace App\Http\Controllers\API;
 use App\Models\PurchaseItem;
 use App\Models\PurchaseInvoice;
 use App\Http\Controllers\Controller;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
+use App\Http\Resources\PurchaseInvoice\PurchaseInvoiceResource;
 use App\Http\Requests\PurchaseInvoice\PurchaseInvoiceCreateRequest;
 use App\Http\Requests\PurchaseInvoice\PurchaseInvoiceUpdateRequest;
-use App\Http\Resources\PurchaseInvoice\PurchaseInvoiceResource;
 use App\Http\Resources\PurchaseInvoice\PurchaseInvoiceWithItemsResource;
 
 class PurchaseInvoiceController extends Controller {
@@ -17,9 +19,13 @@ class PurchaseInvoiceController extends Controller {
      * @return \Illuminate\Http\Response
      */
     public function index() {
+
+        if(!Gate::allows('purchaseInvoice.view')) {
+            abort(403, 'Sorry, permission denied!');
+        }
         
         return PurchaseInvoiceResource::collection(
-            PurchaseInvoice::paginate(10)
+            PurchaseInvoice::latest()->get()
         );
     }
 
@@ -31,11 +37,17 @@ class PurchaseInvoiceController extends Controller {
      */
     public function store(PurchaseInvoiceCreateRequest $request) {
 
+        if(!Gate::allows('purchaseInvoice.create')) {
+            abort(403, 'Sorry, permission denied!');
+        }
+
         $purchaseInvoice = PurchaseInvoice::create([
             'invoice_number' => $request->invoice_number,
-            'invoice_date' => $request->invoice_number,
+            'invoice_date' => $request->invoice_date,
             'supplier_id' => $request->supplier_id,
-            'user_id' => $request->user_id,
+            'invoice_total' => $request->invoice_total,
+            'paid_amount' => $request->paid_amount,
+            'user_id' => Auth::id(),
             'notes' => $request->notes
         ]);
 
@@ -62,8 +74,13 @@ class PurchaseInvoiceController extends Controller {
      * @return \Illuminate\Http\Response
      */
     public function update(
+        
         PurchaseInvoiceUpdateRequest $request, 
         PurchaseInvoice $purchaseInvoice) {
+
+        if(!Gate::allows('purchaseInvoice.update', $purchaseInvoice)) {
+            abort(403, 'Sorry, permission denied!');
+        }
 
         $purchaseInvoice->update($request->all());
         $purchaseInvoice->items()->delete();
